Make the tree pass its own clang-tidy again
Five headers carried diagnostics the config treats as errors, so every commit
touching a file that includes them failed the pre-commit hook whatever it
changed. None of them is a bug; all five are the checker asking for the intent
to be written down.

  - CameraController.h took its deltas as double and assigned them to float
    members eleven times over. The conversions were always meant, so they say
    so now, and the frame delta becomes one float at the top of update().
  - Camera.h computed a half-angle in radians through M_PI, which is double.
    glm::radians is the same number in float.
  - texture_compress.h decoded a normal map's bytes into float without saying
    it was doing that.
  - material_params.h's MaterialType has a numbered ABI the shaders share and a
    COUNT sentinel that deliberately has no number; cert-int09-c wants all or
    none, so the reason is a comment and a NOLINT rather than a sixth number
    nothing may use.
  - env.h reads the environment with std::getenv, which is mt-unsafe only if
    something writes it. Nothing in this tree does, and every read happens while
    the process is still starting up.

+++ src/render/metal/texture_compress.h
@@ -205,9 +205,9 @@ inline void normalizeNormalMap(uint8_t* rgba, int width, int height)
     for (size_t i = 0; i < count; ++i)
     {
         uint8_t* p = rgba + i * 4;
-        const float x = p[0] / 255.0f * 2.0f - 1.0f;
-        const float y = p[1] / 255.0f * 2.0f - 1.0f;
-        const float z = p[2] / 255.0f * 2.0f - 1.0f;
+        const float x = static_cast<float>(p[0]) / 255.0f * 2.0f - 1.0f;
+        const float y = static_cast<float>(p[1]) / 255.0f * 2.0f - 1.0f;
+        const float z = static_cast<float>(p[2]) / 255.0f * 2.0f - 1.0f;
         const float len = std::sqrt(x * x + y * y + z * z);
         if (len < 1e-6f)
         {
